requirements_extra.txt
crfm-helm[together]
crfm-helm[metrics]
google-api-python-client==2.173.0
tiktoken==0.9.0
openai==1.93.2
together==1.5.21
git+https://github.com/mailong25/langfun.git@custom-longform-changes#egg=langfun
rank-bm25==0.2.2
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl#sha256=86cc141f63942d4b2c5fcee06630fd6f904788d2f0ab005cce45aadb8fb73889